Last commit after refactoring

+++ src/main/java/com/begliak/cryptocurrency/entity/CryptoCurrency.java
@@ -23,5 +23,5 @@ public class CryptoCurrency {
     private String symbol;
 
     @Column(name = "price_usd")
-    private Double priceUsd;
+    private BigDecimal priceUsd;
 }
